Move billing-related classes into their own package
