+++ app/src/main/java/me/ccrama/redditslide/util/billing/Base64.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package me.ccrama.redditslide.util;
+package me.ccrama.redditslide.util.billing;
 
 // This code was converted from code at http://iharder.sourceforge.net/base64/
 // Lots of extraneous features were removed.
