+++ app/src/main/java/me/ccrama/redditslide/util/billing/Base64DecoderException.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package me.ccrama.redditslide.util;
+package me.ccrama.redditslide.util.billing;
 
 /**
  * Exception thrown when encountering an invalid Base64 input character.
