+++ app/src/main/java/me/ccrama/redditslide/util/billing/Inventory.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package me.ccrama.redditslide.util;
+package me.ccrama.redditslide.util.billing;
 
 import java.util.ArrayList;
 import java.util.HashMap;
