+++ app/src/main/java/me/ccrama/redditslide/util/billing/Security.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package me.ccrama.redditslide.util;
+package me.ccrama.redditslide.util.billing;
 
 import android.text.TextUtils;
 import android.util.Log;
